+++ client/pages/master/product-list.js
@@ -214,11 +214,6 @@ class ProductList extends localize(i18next)(PageView) {
           })}) {
             items {
               id
-              bizplace {
-                id
-                name
-                description
-              }
               name
               description
               type
@@ -247,7 +242,7 @@ class ProductList extends localize(i18next)(PageView) {
     const response = await client.query({
       query: gql`
           mutation {
-            updateMultipleCompany(${gqlBuilder.buildArgs({
+            updateMultipleProduct(${gqlBuilder.buildArgs({
               patches
             })}) {
               name
@@ -329,25 +324,28 @@ class ProductList extends localize(i18next)(PageView) {
   }
 
   async _deleteProducts() {
-    const names = this.dataGrist.selected.map(record => record.name)
-    if (names && names.length > 0) {
-      const response = await client.query({
-        query: gql`
+    let confirmDelete = confirm('Are you sure?')
+    if (confirmDelete) {
+      const names = this.dataGrist.selected.map(record => record.name)
+      if (names && names.length > 0) {
+        const response = await client.query({
+          query: gql`
             mutation {
               deleteProducts(${gqlBuilder.buildArgs({ names })})
             }
           `
-      })
+        })
 
-      if (!response.errors) {
-        this.dataGrist.fetch()
-        document.dispatchEvent(
-          new CustomEvent('notify', {
-            detail: {
-              message: i18next.t('text.data_updated_successfully')
-            }
-          })
-        )
+        if (!response.errors) {
+          this.dataGrist.fetch()
+          document.dispatchEvent(
+            new CustomEvent('notify', {
+              detail: {
+                message: i18next.t('text.data_updated_successfully')
+              }
+            })
+          )
+        }
       }
     }
   }
