+++ client/pages/master/product-option-list.js
@@ -234,7 +234,7 @@ class ProductOptionList extends connect(store)(localize(i18next)(PageView)) {
     const response = await client.query({
       query: gql`
           mutation {
-            updateMultipleCompany(${gqlBuilder.buildArgs({
+            updateMultipleProductOption(${gqlBuilder.buildArgs({
               patches
             })}) {
               name
@@ -329,10 +329,12 @@ class ProductOptionList extends connect(store)(localize(i18next)(PageView)) {
   }
 
   async _deleteProductOptions() {
-    const names = this.dataGrist.selected.map(record => record.name)
-    if (names && names.length > 0) {
-      const response = await client.query({
-        query: gql`
+    let confirmDelete = confirm('Are you sure?')
+    if (confirmDelete) {
+      const names = this.dataGrist.selected.map(record => record.name)
+      if (names && names.length > 0) {
+        const response = await client.query({
+          query: gql`
             mutation {
               deleteProductOptions(${gqlBuilder.buildArgs({
                 product: {
@@ -342,17 +344,18 @@ class ProductOptionList extends connect(store)(localize(i18next)(PageView)) {
               })})
             }
           `
-      })
+        })
 
-      if (!response.errors) {
-        this.dataGrist.fetch()
-        document.dispatchEvent(
-          new CustomEvent('notify', {
-            detail: {
-              message: i18next.t('text.data_updated_successfully')
-            }
-          })
-        )
+        if (!response.errors) {
+          this.dataGrist.fetch()
+          document.dispatchEvent(
+            new CustomEvent('notify', {
+              detail: {
+                message: i18next.t('text.data_updated_successfully')
+              }
+            })
+          )
+        }
       }
     }
   }
